backend/requirements.txt
google-genai==1.70.0
feedparser==6.0.11
python-dotenv==1.0.1
httpx==0.28.1
beautifulsoup4==4.12.3
